requirements.txt
# Core dependencies
gradio>=4.0.0
numpy>=1.24.0
pandas>=2.0.0

# Video processing
opencv-python>=4.8.0
ffmpeg-python>=0.2.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
scipy>=1.11.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0

# HTTP and networking
requests>=2.31.0
urllib3>=2.0.0

# Stream extraction (required for Kick.com)
yt-dlp>=2023.0.0

# Utilities
python-dateutil>=2.8.0
